added http-equiv type
